# apps/api/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type AuditEventType string

const (
	AuditEventTypeReservation AuditEventType = "reservation"
	AuditEventTypePricing     AuditEventType = "pricing"
	AuditEventTypeReview      AuditEventType = "review"
	AuditEventTypeOwner       AuditEventType = "owner"
	AuditEventTypeSystem      AuditEventType = "system"
)

func (e *AuditEventType) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = AuditEventType(s)
	case string:
		*e = AuditEventType(s)
	default:
		return fmt.Errorf("unsupported scan type for AuditEventType: %T", src)
	}
	return nil
}

type NullAuditEventType struct {
	AuditEventType AuditEventType
	Valid          bool // Valid is true if AuditEventType is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullAuditEventType) Scan(value interface{}) error {
	if value == nil {
		ns.AuditEventType, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.AuditEventType.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullAuditEventType) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.AuditEventType), nil
}

type BookingStatus string

const (
	BookingStatusPending   BookingStatus = "pending"
	BookingStatusApproved  BookingStatus = "approved"
	BookingStatusRejected  BookingStatus = "rejected"
	BookingStatusCancelled BookingStatus = "cancelled"
	BookingStatusPaid      BookingStatus = "paid"
)

func (e *BookingStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = BookingStatus(s)
	case string:
		*e = BookingStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for BookingStatus: %T", src)
	}
	return nil
}

type NullBookingStatus struct {
	BookingStatus BookingStatus
	Valid         bool // Valid is true if BookingStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullBookingStatus) Scan(value interface{}) error {
	if value == nil {
		ns.BookingStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.BookingStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullBookingStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.BookingStatus), nil
}

type ReviewStatus string

const (
	ReviewStatusPending  ReviewStatus = "pending"
	ReviewStatusApproved ReviewStatus = "approved"
	ReviewStatusRejected ReviewStatus = "rejected"
)

func (e *ReviewStatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = ReviewStatus(s)
	case string:
		*e = ReviewStatus(s)
	default:
		return fmt.Errorf("unsupported scan type for ReviewStatus: %T", src)
	}
	return nil
}

type NullReviewStatus struct {
	ReviewStatus ReviewStatus
	Valid        bool // Valid is true if ReviewStatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullReviewStatus) Scan(value interface{}) error {
	if value == nil {
		ns.ReviewStatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.ReviewStatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullReviewStatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.ReviewStatus), nil
}

type AdminUser struct {
	ID           pgtype.UUID
	Email        string
	PasswordHash string
	CreatedAt    pgtype.Timestamptz
}

type AuditEvent struct {
	ID         pgtype.UUID
	VillaSlug  string
	Type       AuditEventType
	Message    string
	ActorEmail string
	CreatedAt  pgtype.Timestamptz
}

type Booking struct {
	ID         pgtype.UUID
	VillaSlug  string
	GuestName  string
	GuestEmail string
	GuestPhone string
	CheckIn    pgtype.Date
	CheckOut   pgtype.Date
	Adults     int16
	Children   int16
	Message    string
	Status     BookingStatus
	CreatedAt  pgtype.Timestamptz
	UpdatedAt  pgtype.Timestamptz
	Source     string
}

type PriceOverride struct {
	VillaSlug  string
	Date       pgtype.Date
	PriceCents int32
	CreatedAt  pgtype.Timestamptz
	UpdatedAt  pgtype.Timestamptz
}

type Review struct {
	ID                pgtype.UUID
	BookingID         pgtype.UUID
	VillaSlug         string
	AuthorName        string
	Rating            int16
	Body              string
	Status            ReviewStatus
	CreatedAt         pgtype.Timestamptz
	UpdatedAt         pgtype.Timestamptz
	Featured          bool
	Meta              string
	Source            string
	RatingCleanliness pgtype.Numeric
	RatingComfort     pgtype.Numeric
	RatingLocation    pgtype.Numeric
	RatingHost        pgtype.Numeric
	RatingValue       pgtype.Numeric
}

type SeasonRule struct {
	ID         pgtype.UUID
	VillaSlug  string
	Label      string
	StartDate  pgtype.Date
	EndDate    pgtype.Date
	PriceCents int32
	CreatedAt  pgtype.Timestamptz
	UpdatedAt  pgtype.Timestamptz
}

type VillaPricingSetting struct {
	VillaSlug         string
	BasePriceCents    int32
	MinNights         int16
	CleaningFeeCents  int32
	ConciergeFeeCents int32
	CreatedAt         pgtype.Timestamptz
	UpdatedAt         pgtype.Timestamptz
}