# apps/api/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Casa Dana API
  version: 0.1.0
  description: |
    Casa Dana villa rental API. Public endpoints for booking submission and
    availability lookup. Villa content (names, photos, descriptions) is static
    in the frontend and not exposed by this API.
servers:
  - url: http://localhost:8080
    description: local dev
  - url: https://casa-dana.com
    description: production
...
paths:
...
  /api/reviews:
    post:
      operationId: submitReview
      tags: [reviews]
      summary: Submit a review for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReviewRequest"
      responses:
        "201":
          description: Review accepted (pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Booking not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: A review already exists for this booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Hard delete a review
      security:
        - adminSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "401":
          description: Missing or invalid admin session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Review not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      operationId: patchReview
      tags: [reviews]
      summary: Moderate a review
      description: |
        Partial update of status, featured flag, or editorial fields. Omitted
        fields are left unchanged.
      security:
        - adminSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchReviewRequest"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }
        "401":
          description: Missing or invalid admin session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Review not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/admin/reviews:
    get:
      operationId: listAdminReviews
      tags: [reviews]
      summary: List reviews for moderation
      description: Unlike the public list, this returns every status.
      security:
        - adminSession: []
      parameters:
        - name: villa_slug
          in: query
          required: true
          schema: { type: string }
          example: casadana
        - name: status
          in: query
          description: Omit to return all statuses.
          schema: { $ref: "#/components/schemas/ReviewStatus" }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListReviewsResponse" }
        "401":
          description: Missing or invalid admin session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Missing or unknown villa_slug
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: createAdminReview
      tags: [reviews]
      summary: Create a review without a booking
      description: |
        For testimonials collected outside the booking flow. Guest-submitted
        reviews still go through `POST /api/reviews`.
      security:
        - adminSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminReviewRequest"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }
        "401":
          description: Missing or invalid admin session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/villas/{slug}/reviews:
    get:
      operationId: listVillaReviews
      tags: [reviews]
      summary: List published reviews for a villa
      description: |
        Public endpoint — returns only `approved` reviews. Use
        `GET /api/admin/reviews` to see pending and rejected ones.
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListReviewsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/villas/{slug}/reviews/meta:
    get:
      operationId: getVillaReviewMeta
      tags: [reviews]
      summary: Published rating figures for a villa
      description: |
        Computed from the villa's approved reviews. Read-only: the figures move
        when a review is added, moderated or removed, never by being set.
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Review meta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewMeta" }
        "404":
          description: Villa slug not in allowlist
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: Internal server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  schemas:
...
    ReviewStatus:
      type: string
      enum: [pending, approved, rejected]
    Review:
      type: object
      required:
        - id
        - booking_id
        - villa_slug
        - author_name
        - rating
        - body
        - status
        - meta
        - source
        - featured
        - created_at
      properties:
        id:          { type: string, format: uuid }
        booking_id:
          type: string
          format: uuid
          nullable: true
          description: Null for admin-authored reviews, which have no booking behind them.
        villa_slug:  { type: string }
        author_name: { type: string }
        rating:      { type: integer, minimum: 1, maximum: 5 }
        body:        { type: string }
        status:      { $ref: "#/components/schemas/ReviewStatus" }
        meta:
          type: string
          description: Free-text context line shown under the author name.
          example: Paris, France · Stayed June 2026
        source:
          type: string
          description: Where the review came from.
          example: via Airbnb · Couple
        featured:
          type: boolean
          description: Pinned to the top of the public reviews section.
        categories:  { $ref: "#/components/schemas/ReviewCategoryRatings" }
        created_at:  { type: string, format: date-time }
    ReviewCategoryRatings:
      type: object
      description: |
        Per-category scores for a single review. Every field is optional and
        nullable: a null means the reviewer did not score that category, which
        keeps it out of the villa's average for that category rather than
        pulling it down with a zero.
      properties:
        cleanliness: { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
        comfort:     { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
        location:    { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
        host:        { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
        value:       { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
    CreateAdminReviewRequest:
      type: object
      required: [villa_slug, author_name, rating]
      properties:
        villa_slug:  { type: string, minLength: 1, maxLength: 64 }
        author_name: { type: string, minLength: 1, maxLength: 120 }
        rating:      { type: integer, minimum: 1, maximum: 5 }
        body:        { type: string, maxLength: 2000, default: "" }
        status:      { $ref: "#/components/schemas/ReviewStatus" }
        meta:        { type: string, maxLength: 2000, default: "" }
        source:      { type: string, maxLength: 64, default: "" }
        featured:    { type: boolean, default: false }
        categories:  { $ref: "#/components/schemas/ReviewCategoryRatings" }
    PatchReviewRequest:
      type: object
      description: Every field is optional; an omitted field is left unchanged.
      properties:
        status:     { $ref: "#/components/schemas/ReviewStatus" }
        featured:   { type: boolean }
        meta:       { type: string, maxLength: 2000 }
        source:     { type: string, maxLength: 64 }
        body:       { type: string, maxLength: 2000 }
        rating:     { type: integer, minimum: 1, maximum: 5 }
        categories: { $ref: "#/components/schemas/ReviewCategoryRatings" }
    ReviewBreakdown:
      type: object
      required: [cleanliness, comfort, location, host, value]
      description: |
        Per-category averages over the villa's approved reviews. A null means no
        approved review has scored that category yet, so the bar is left off
        rather than drawn at zero.
      properties:
        cleanliness: { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
        comfort:     { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
        location:    { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
        host:        { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
        value:       { type: number, format: double, minimum: 1, maximum: 5, nullable: true }
    ReviewMeta:
      type: object
      required: [display_avg, display_count, breakdown]
      description: |
        The villa's published rating, computed from its approved reviews.
        Nothing here is stored or hand-entered: moderating a review into or out
        of `approved` is what moves these numbers, so the figures a guest reads
        always match the reviews on show. `display_avg` is 0 when the villa has
        no approved reviews — read `display_count` to tell that apart from a
        genuine average.
      properties:
        display_avg:   { type: number, format: double, minimum: 0, maximum: 5, example: 4.92 }
        display_count: { type: integer, minimum: 0, example: 128 }
        breakdown:     { $ref: "#/components/schemas/ReviewBreakdown" }
    ListReviewsResponse:
      type: object
      required: [reviews]
      properties:
        reviews:
          type: array
          items: { $ref: "#/components/schemas/Review" }
    SubmitReviewRequest:
      type: object
      required: [booking_id, author_name, rating]
      properties:
        booking_id:  { type: string, format: uuid }
        author_name: { type: string, minLength: 1, maxLength: 120 }
        rating:      { type: integer, minimum: 1, maximum: 5 }
        body:        { type: string, maxLength: 2000, default: "" }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Machine-readable error code. Known codes:
                `DATES_CONFLICT`, `UNKNOWN_VILLA`, `VALIDATION`,
                `INTERNAL`, `INVALID_STATUS`.
              example: DATES_CONFLICT
            message:
              type: string
              example: those dates are not available
...
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: casa_admin_session