# packages/api/src/generated/schemas/createAdminReviewRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Casa Dana API
 * Casa Dana villa rental API. Public endpoints for booking submission and
availability lookup. Villa content (names, photos, descriptions) is static
in the frontend and not exposed by this API.

 * OpenAPI spec version: 0.1.0
 */
import type { ReviewStatus } from './reviewStatus';
import type { ReviewCategoryRatings } from './reviewCategoryRatings';

export interface CreateAdminReviewRequest {
  /**
   * @minLength 1
   * @maxLength 64
   */
  villa_slug: string;
  /**
   * @minLength 1
   * @maxLength 120
   */
  author_name: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** @maxLength 2000 */
  body?: string;
  status?: ReviewStatus;
  /** @maxLength 2000 */
  meta?: string;
  /** @maxLength 64 */
  source?: string;
  featured?: boolean;
  categories?: ReviewCategoryRatings;
}