# packages/api/src/generated/schemas/patchReviewRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Casa Dana API
 * Casa Dana villa rental API. Public endpoints for booking submission and
availability lookup. Villa content (names, photos, descriptions) is static
in the frontend and not exposed by this API.

 * OpenAPI spec version: 0.1.0
 */
import type { ReviewStatus } from './reviewStatus';
import type { ReviewCategoryRatings } from './reviewCategoryRatings';

/**
 * Every field is optional; an omitted field is left unchanged.
 */
export interface PatchReviewRequest {
  status?: ReviewStatus;
  featured?: boolean;
  /** @maxLength 2000 */
  meta?: string;
  /** @maxLength 64 */
  source?: string;
  /** @maxLength 2000 */
  body?: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating?: number;
  categories?: ReviewCategoryRatings;
}