# packages/api/src/generated/schemas/review.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Casa Dana API
 * Casa Dana villa rental API. Public endpoints for booking submission and
availability lookup. Villa content (names, photos, descriptions) is static
in the frontend and not exposed by this API.

 * OpenAPI spec version: 0.1.0
 */
import type { ReviewStatus } from './reviewStatus';
import type { ReviewCategoryRatings } from './reviewCategoryRatings';

export interface Review {
  id: string;
  /**
   * Null for admin-authored reviews, which have no booking behind them.
   * @nullable
   */
  booking_id: string | null;
  villa_slug: string;
  author_name: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  body: string;
  status: ReviewStatus;
  /** Free-text context line shown under the author name. */
  meta: string;
  /** Where the review came from. */
  source: string;
  /** Pinned to the top of the public reviews section. */
  featured: boolean;
  categories?: ReviewCategoryRatings;
  created_at: string;
}