# packages/api/src/generated/schemas/reviewCategoryRatings.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Casa Dana API
 * Casa Dana villa rental API. Public endpoints for booking submission and
availability lookup. Villa content (names, photos, descriptions) is static
in the frontend and not exposed by this API.

 * OpenAPI spec version: 0.1.0
 */

/**
 * Per-category scores for a single review. Every field is optional and
nullable: a null means the reviewer did not score that category, which
keeps it out of the villa's average for that category rather than
pulling it down with a zero.

 */
export interface ReviewCategoryRatings {
  /**
   * @minimum 1
   * @maximum 5
   * @nullable
   */
  cleanliness?: number | null;
  /**
   * @minimum 1
   * @maximum 5
   * @nullable
   */
  comfort?: number | null;
  /**
   * @minimum 1
   * @maximum 5
   * @nullable
   */
  location?: number | null;
  /**
   * @minimum 1
   * @maximum 5
   * @nullable
   */
  host?: number | null;
  /**
   * @minimum 1
   * @maximum 5
   * @nullable
   */
  value?: number | null;
}