# packages/api/src/generated/schemas/reviewMeta.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Casa Dana API
 * Casa Dana villa rental API. Public endpoints for booking submission and
availability lookup. Villa content (names, photos, descriptions) is static
in the frontend and not exposed by this API.

 * OpenAPI spec version: 0.1.0
 */
import type { ReviewBreakdown } from './reviewBreakdown';

/**
 * The villa's published rating, computed from its approved reviews.
Nothing here is stored or hand-entered: moderating a review into or out
of `approved` is what moves these numbers, so the figures a guest reads
always match the reviews on show. `display_avg` is 0 when the villa has
no approved reviews — read `display_count` to tell that apart from a
genuine average.

 */
export interface ReviewMeta {
  /**
   * @minimum 0
   * @maximum 5
   */
  display_avg: number;
  /** @minimum 0 */
  display_count: number;
  breakdown: ReviewBreakdown;
}